+++ packages/nimble-components/src/table/components/row/index.ts
@@ -133,6 +133,12 @@ export class TableRow<
     @observable
     public animationClass = '';
 
+    /**
+     * @internal
+     */
+    @attr({ attribute: 'allow-hover', mode: 'boolean' })
+    public allowHover = false;
+
     @volatile
     public get isTopLevelParentRow(): boolean {
         return this.isParentRow && this.nestingLevel === 0;